Pass Pages type into Zig
I've been considering two options for building out this platform:

1. Implement as much as possible of in the platform (Roc) code. Add a
   bunch of internal-only calls passing bits of data between Roc and
   Zig. Upsides: I can avoid passing custom types between Roc and Zig,
   working around not yet existing. Don't need to figure out how to call
   lambda's from Zig.

2. Implement as much as possible in host (Zig) code. Implement just the
   copy! and bootstrap! effects. Upsides: way fewer interactions between
   platform and host should lead to simpler code. Can make use of zig
   standard library for path manipulation, given Roc's is still a WIP.

In the end I decided (2) is the better long-term choice, and it's worth
investing in figuring out how to do it. So this commit changes the
limited code I have towards that scenario, and sets up some code to
generate type descriptions using glue so I can manually implement the
related types in Zig.