Put path information into arraylists
Part of the motivation here is that it should be possible to further
grow array lists later, as a user might add additional source files
after the initial run.

Another is that having all these be arraylists sets us up for file
generation being easier, for when we have different processing
strategies in a bit.