Change API for metadata exchange
One feature I'd like to add next is letting pages depend on the metadata of
other pages. That will enable things creating lists of all the posts, adding
links to next/previous pages, and more. I'd like to do this in such a way that
Jay can keep a dependency graph of which pages depend on which other pages, so
that when a change happens only pages dependent on that change will need to be
reorganized.

I think the dependency tracking is easiest to realize by using effects for it,
in which I want to avoid using effects for defining page groups. Otherwise
there would be two contexts in which different sets of effects are usable, and
when folks use an effect in the wrong context I'd have to throw an error.