Extract non-Roc interop code from main.zig
The main.zig file needs to manage some interaction with the Roc program,
and in the future might need to use some globals for calls from Roc into
Zig to act on program data. Lets keep all that in `main.roc`, and
extract most of the logic into lib.zig.