Move .pages of .rules onto site directly
This simplifies logic in a couple of places, among them the web server
implementation we want to do next.
It's also possible that in the future we might transform output paths,
for instance to hash source contents, which would make it hard for
people to know what pattern to match on.