Improve bootstrapping and memory use
We now limit use of the Site arena to just the data stored in the Site
object, and avoid doing temporary allocations in that arena. Currently
this doesn't really make a difference because the process exits after
generating a site anyway, but in upcoming changes we're looking to have
it continue running and respond to changes in source files.

Bootstrapping now also populates the `pages` field on the `Site`, so we
can immediately continue with generating a site after bootstrapping.