Vendor Zig fanotify stdlib code
Zig added a watch mode for its own build system. It's merged in master
but not yet part of a stable release. The implementation is a good
example of how I might go about implementing filewatching myself:

https://ziglang.org/documentation/master/std/#src/std/Build/Watch.zig

Possibly because of this work the current master version of Zig has some
extra conveniences for the Linux fanotify system. This commits picks
those functions from the Zig source code and vendors them in the
project, until the time that 0.14 becomes available and I will upgrade
to that.