Add path interning helper
I'm trying to pull some dedicated reponsibilities out of Site, because
it's growing a bit beyind my ability to comprehend. I'm trying to cut in
such a way that data with the same lifetime stays together.

I think it's okay to intern source paths and not free these until the
program ends. Source paths will typically remain pretty stable unless
the site author does a weird amount of renaming. If I want to optimize
that use case I might add functionality to re-intern all paths if the
amount of storage we have for paths exceeds the amount we needed for the
initial source files by some factor, but I think it's probably overkill.