Intern paths
For the watch mode it helps to have some way to uniquely refer to paths
that preferably isn't a bit array of bytes. I had PageIndex for this
purpose, but that was owned by the Site and I don't want to create
dependencies from the watcher on the site or vice-versa.

This big refactor interns a separate helper for interning paths. Each
interned path gets a unique index that can be used by the site. The site
keeps a record of pages by path index. It can use a single datastructure
for efficient lookups by the source path, output path, or web path.