Use record builder for combining pages
The previous API of combining routes by passing a list doesn't
generalize well. It requires all the different page types to unify to a
single type, which worked with the examples I've used so far but won't
generally always be the case: someone might want to mix page types.

Using record builders in this way avoids needing to introduce another
types next to 'Pages' that would be tricky to name.