Manually write payload for interned strings
Previously interned paths all had a unique index, but if we allow this
index to be treated as a bit of custom payload, then we can directly
write the page index there and avoid some indirection.

This prepares the path for intering not just paths, but other strings as
well.