Extract Site#ensurePage
I'd like to separate the step of initializing a page from a source path
and some rules, from the step of of scanning the source file and
obtaining relevant data from that.

The information from both steps needs to be kept for different amount of
times, so I might split the `Page` type and keep them separate to make
memory management easier.