Perform page scanning immediately
We were queueing up page scanning, like we did page generation. This
combines these steps into two. Giving it more thought, I can justify
generation happening after all pages are scanned, but scanning itself
can be done as soon as we generate the page, and in fact not doing it is
causing some difficulty that tests are helping to find.