Don't use gpa as a global
Doing this was hiding a couple of bugs and making it impossible to use
gpa as a variable in a function (because of shadowing).

This removes the optimization for the moment where we avoid regenerating
dependent files when the frontmatter of the dependency doesn't change.
This optimization wasn't working properly, it was doing a pointer
comparison between the old and new metadata pointers, but the metadata
might end up in the same memory location and so not receive a new
pointer. I might bring this back in the future, implemented in a better
way.