Use cache directory as default output directory
This is intended to be used as the output directory when running jay in
dev/watch mode. I want to add, in a separate change, the option to pass
an output directory as a command line param, but that should run jay in
production mode and not start the watcher or webserver.