Serve clients in parallel
Ideally I don't spawn a thread per client, but I think if I want to do
something about that I need to pull in a pretty heavy event-loopy
dependency, which I'd like to avoid. The server is only for local
development, so realistically only a few clients will connect anyway
(maybe a couple of different browsers).