Don't accidentally `rm -r /`
Before this commit we would wipe the output directory passed with a
`prod` command, no questions asked. This made it easy to make
catastrophic errors.

In the new setup we'll always generate output in a sub-directory with a
known name. That's mighg be a bit less convenient when running a
production build, but it allows us to just wipe a directory with that
specific name.