Fix deadlock when page has dependency on self
The page mutex would be locked as we started generating the page,
rightly because we don't want the web server to serve the page while
we're doing that. If the page has a dependency on itself it will try to
get the lock again when it reads its own metadata. This caused the
deadlock.

I think I'm locking more than I need to. The only danger is that a
modification on the page on the main thread overlaps with read a page on
a web thread. I don't need to lock for reads on the main thread because
there's no other thread that might be modifying the page at the same
time.

I should take a closer look at my use of mutexes more broadly.

This also removes the `page.scanned.deleted` property. We can check if a
page exists by verifying `page.scanned` is not null instead, making the
boolean unnecessary.