Generate absolute paths on pages
I'm still figuring out the best solution here. Using relative paths by
default is definitely wrong and gives when navigating to a page from
anywhere but the home page. Absolute paths mean that the generated site
expects to be hosted on the root of a domain. In the future I might want
to take the domain name and use that, but I have to think about it more.