Drop tree-sitter-highlight
It's a rust project, and including it means part of the build runs
outside of Zig's control, which results in a loss of cross-compilation
ability.

tree-sitter-highlight is a cool tool, but it's small enough that writing
a Zig implementation isn't too much effort. As I finished integrating it
it dawned on me that the integration effort might have been better spent
writing the logic I need in Zig.

In addition, a significant portion of the tree-sitter-highlight
implementation is dedicated to implementing nested highlighting, where
one language is embedded in another. While cool it's not a super high
priority for me, and I'm fine with Jay not having it for now.

Finally tree-sitter-highlight requires the user to specify which
highlight groups they are interested in. For Jay I'd like to get all of
them, and implementing this required a hacky parse of the highlights.scm
query to find all support highlight groups. Writing a custom
implementation removes the need for this logic.

This highlighting implementation does not yet ensure that parameters are
colored consistently, the way that tree-sitter-highlighter does. This
might be something to add in the future. Other than that the reported
highlights are slightly different, maybe even a bit better on the whole.
I think this might be to do with what both highlighting implementations
do when encountering a node that's matched by two different highlight
patterns. The implementation can be tweaked more, but seems passable for
the moment.