Split build into separate build types
The goal here is to separate out the release build so it can run across
multiple platforms. Certain parts of the build, such as building the
docs site, don't need to run in multiple platforms, can't even run on
multiple platforms because even if we can build binaries for multiple
targets on a single machine, that machine will only be able to run once
of those binaries.