Add examples-based integration tests
The idea is that Zig will run production builds for each project in the
examples/ directory, and test that the generated files match the ones
committed.

This will not exercise the development-server aspects of the code. I've
some coverage over file watching from regular zig tests, but not much
for the webserver yet. We'll see how to best approach those later.